Improved Wave and Turret Behavior, added Music Manager
Music Manager will (hopefully) preload and play tracks when
appropriate. The currently used track is TEMPORARY and will not be in
the final release. It is for DEMO PURPOSES ONLY.

Wave behavior now uses the same timing system as Jolene’s falling speed
and the wall animator, making it easily possible to have waves “hold
position” on the wall.

Turrets now fire at a slightly different rate.